Exclusions are applied to all environments
There's currently no note about the fact that one you create an exclusion, it's applied to all environments. 

There's a very ambiguous note at the bottom that says: 

"Be aware that exclusions are available across the entire account. Ensure that you consider this before deleting an exclusion. Deleted exclusions cannot be recovered." 

However, "Available" doesn't mean "applied to". The exclusions can be available to anyone to edit or see in the account, but that doesn't mean they are applied to all environments.