KB-2062 Bennet Roy Joseph suggestion
Added:

A visitor's geo information is determined from the originating IP address of a [!DNL Target] location request (mbox request). The IP-to-geo resolution is done for the first call of a new session. This means, if the IP address of a visitor changes mid session of a visit, the geo information is still based on the IP address of the first call.