Update working-with-customer-attributes.md
I made some changes in the Experience Cloud interface and admin guide, so I made a few edits on this page to mirror those changes. If you refer to "Customer Attributes" (the feature), it's title case. When generally referring to customer attribute data, it's lowercase. The feature name is plural, so I capitalize it in that situation. (Knowing when we're talking about CA the feature vs. generic "customer attribute data" is a little tricky.)